+++ ui-src/js/header-controller.js
@@ -25,18 +25,26 @@ function HeaderController($scope, $state, $http, growl, HydraAuthService, Config
         }
     }
 
-    $scope.$on("user:loggedIn", function (event, data) {
+    function onLogin(data) {
         $scope.showAdmin = data.maySeeAdmin;
         $scope.showStats = data.maySeeStats;
         $scope.showLoginout = true;
         $scope.loginlogoutText = "Logout";
+    }
+
+    $scope.$on("user:loggedIn", function (event, data) {
+        onLogin(data);
     });
 
-    $scope.$on("user:loggedOut", function (event, data) {
+    function onLogout() {
         $scope.showAdmin = !bootstrapped.adminRestricted;
         $scope.showStats = !bootstrapped.statsRestricted;
         $scope.loginlogoutText = "Login";
         $scope.showLoginout = bootstrapped.adminRestricted || bootstrapped.statsRestricted || bootstrapped.searchRestricted;
+    }
+
+    $scope.$on("user:loggedOut", function (event, data) {
+        onLogout();
     });
 
     $scope.loginout = function () {
@@ -48,7 +56,8 @@ function HeaderController($scope, $state, $http, growl, HydraAuthService, Config
             }
             else if (ConfigService.getSafe().authType == "form") {
                 growl.info("Logged out");
-            } 
+            }
+            onLogout();
             $state.go("root.search");
         } else {
             if (ConfigService.getSafe().authType == "basic") {
@@ -60,6 +69,7 @@ function HeaderController($scope, $state, $http, growl, HydraAuthService, Config
                 } 
                 $http.get("internalapi/askforpassword", {params: params}).then(function () {
                     growl.info("Login successful!");
+                    //onLogin();
                     $state.go("root.search");
                 })
             } else if (ConfigService.getSafe().authType == "form") {
@@ -67,9 +77,6 @@ function HeaderController($scope, $state, $http, growl, HydraAuthService, Config
             } else {
                 growl.info("You shouldn't need to login but here you go!");
             }
-
         }
-
     }
-
 }