+++ changelog.md
@@ -1,6 +1,11 @@
 # NZB Hydra changelog
 
 ----------
+### 0.2.108
+Changed: Rewrote auth handling. Unfortunately form based auth only works when calls from the GUI are done, as soon as you call any function from outside (e.g. CP) the token header is
+ missing and you will be asked for credentials using basic auth. Now, when you enter them, hydra will accept them even with form based auth enabled.
+  I might change that behavior in the future but for now it stands.
+
 ### 0.2.107
 Fixed: Repeating search from history wouldn't work (in multiple ways).